requirements.txt
beautifulsoup4==4.4.1
celery==3.1.23
Django==1.9.6
djangorestframework==3.5.3
django-allauth==0.25.2
django-celery-email==1.1.4
django-environ==0.4.0
django-extensions==1.6.7
django-rest-auth-0.9.0


coverage==4.1
django-nose==1.4.3
factory-boy==2.7.0
fake-factory==0.5.7
gunicorn==19.6.0
ipdb==0.10.0
nose-timer==0.6.0
flake8==3.2.1
psycopg2==2.6.1
selenium==2.53.2
whitenoise==3.1
